fix: ensure namespaces will be matched and rewritten correctly
Adds tests to cover an edge case I uncovered in Apigility applications,
whereby the `ZF\Apigility` and `ZF\Apigility\Documentation` modules
listed in `modules.config.php` were not being rewritten correctly.

The solution was to add additional variants of the namespaces with
differences in escape character counts.

This patch also adds the first tests for a test suite, covering the edge
cases.