fix: remove vendor directory
This patch adds logic to remove the configured vendor directory, if it
exists, prior to performing the migration. This ensures:

- it will not rewrite code in the vendor directory
- a clean install will occur after the migration

What I observed was that if ZF components had been installed as
dependencies of other third-party libraries previously, Composer would
detect those, and decide no installation task was required for those
dependencies — which in turn meant the dependency plugin had nothing to
intercept and rewrite.

As part of this fix, I also rewrote some of the internal logic,
performing some "extract method" refactors so I could better understand
the logic.