fix: rewrite the rewrites
The previous rewrite strategy had a small number of very generic
rewrites, and then a growing number of exclusions. While this worked,
users were reporting more and more strings that were problematic, and I
was worried about having a whack-a-mole scenario.

This patch takes the opposite approach: it contains a ton of very
specific rewrites, and very few generic ones (basically none). This
ensures that we only rewrite what we know about _specifically_ instead
of making assumptions and rewriting stuff we do not know about.

Essentially, I grepped through our sources for:

- Package names
- Namespaces
- Script names
- Configuration keys

and, in the latter, I specfically looked for things like the legacy
plugin aliases.

Additionally, I've pushed the new constant definitions to the end of the
class, to make it easier to find the public API.

I have tested this across my various scenarios, and all scenarios
continue to hold at this point.