feat: Migrate zendframework/zendframework package to component packages
Currently, if you try and migrate a project that requires the
"zendframework/zendframework" package, that package is not rewritten; it
is instead retained verbatim.

This patch adds two things:

- A concept of a "special case" for migrations. These are provided the
  filename and the file contents in order to determine if they should
  act; if so, they are then provided the contents on which to perform
  replacements.

- A "ComposerJsonZendFrameworkPackageSpecialCase" that operates on
  `composer.json` files that refer to the "zendframework/zendframework"
  package.

This latter compares the version against a list of version ranges in
order to deliver a list of packages to use instead. When "self.version"
was used to determine which package version to use (generally on
versions less than 2.5.0), it resolves to the latest version of the
package that will fulfill the constraint.

I have provided tests to validate that the correct replacements are
made, and have tested with some local projects to verify that an
installable set of dependencies are available when complete.

Signed-off-by: Matthew Weier O'Phinney <matthew@weierophinney.net>