fix: better testing strategy
Per comments from @michalbundyra, I've refactored the tests to provide
an "output" `composer.json` file to test assertions against. This
ensures that the replacements are done in a way that keeps the JSON
integrity (e.g. no trailing commas; commas after every list item; etc.),
and which also clearly demonstrates what we are looking for when done.

Tests continue to pass, which allows me to move on to refactoring the
special case class itself.

Signed-off-by: Matthew Weier O'Phinney <matthew@weierophinney.net>